==> pegasus2_selection/1_examine_all_asthmatics_aggregated_data.Rmd <==
---
title: "Examining all asthmatics from Biobank"
subtitle: "Using aggregated data generated by Meryl"
author: "MHuang"
date: "11/16/2020"
output:   
  html_document: 
    toc: yes
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}

## Packages

pkg <- c("tidyverse", "data.table", "here", "DT", "readxl", "grid", "gridExtra")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}


## Paths

asthma_dir <- "C:/Users/MHuang/Dropbox (Partners HealthCare)/Asthma/Cleaned_Generated_Files/2020-11-09/"
asthma_rpdr_dir <- "C:/Users/MHuang/Dropbox (Partners HealthCare)/Asthma/RPDR_Downloads/2020-06-30/"
pegasus_dir <- "C:/Users/MHuang/Dropbox (Partners HealthCare)/PEGASUS/rpdr_data/"


## File names

asthma_fname <- str_c(asthma_dir, "RPDR_cleaned_data_2020-11-09.csv")
asthma_bf1st_fname <- str_c(asthma_dir, "RPDR_cleaned_data_2020-11-09_Restrict_Before_First_Collection_Date.csv")
pegasus1_id_fname <- str_c(pegasus_dir, "PEGASUS_Subject_IDs_n1140.csv")
rpdr_file_header <- str_c(asthma_rpdr_dir, "mns43_063020114847776198_")
rpdr_file_ending <- ".txt"
med_map_fname <- str_c(pegasus_dir, "Medication_Mapping.xlsx")

```

# 1. Read in lists and exclude existing PEGASUS subjects

```{r}

asthma <- fread(asthma_fname); nrow(asthma)
pegasus1_id <- fread(pegasus1_id_fname)


## Keep only those with plasma samples

asthma[, .N, Plasma_Existence_Yes_No_To_6_30_2020]
asthma <- asthma[!(Biobank_Subject_ID %in% pegasus1_id$SubjectID) & 
                         Plasma_Existence_Yes_No_To_6_30_2020 == "Yes", ]; nrow(asthma)

asthma[, ':='(asthma_existence = Asthma_current_or_past_history_custom_PPV_0_80PPV_Existence_Yes_No, 
              asthma_ppv = Asthma_current_or_past_history_custom_PPV_0_80PPV_List_of_All_Values, 
              copd_existence = COPD_current_or_past_history_custom_PPV_0_75PPV_Existence_Yes_No, 
              copd_ppv = COPD_current_or_past_history_custom_PPV_0_75PPV_List_of_All_Values)]
asthma[, .N, .(asthma_existence, is.na(asthma_ppv), asthma_ppv > 0.85, asthma_ppv > 0.9, asthma_ppv > 0.95)]

asthma <- asthma[asthma_existence == "Yes", ]; nrow(asthma)
asthma[, .N, .(copd_existence, copd_ppv > 0.9, Any_Other_chronic_obstructive_pulmonary_disease)]
asthma[, .N, .(nCollection_Dates, is.na(All_Collection_Dates))][order(nCollection_Dates)]


## Also trim diagnosis data

dia <- fread(str_c(rpdr_file_header, "Dia", rpdr_file_ending), quote = "")
dia <- dia[EMPI %in% asthma$EMPI, ]
dia[, dia_date := as.Date(Date, format = "%m/%d/%Y")]
dia_condensed <- dia %>% group_by(Diagnosis_Name, Code_Type, Code) %>% summarise() %>% as.data.table()

```

# 2. Comorbidities

```{r}

datatable(dia_condensed, filter = "top")

```

## 2.1. COPD

```{r}

copd_nm <- dia_condensed[grepl("(?i)chronic obstructive lung disease|chronic obstructive pulmonary disease", 
                               Diagnosis_Name), unique(Diagnosis_Name)]
copd_nm
copd_empi <- dia[Diagnosis_Name %in% copd_nm, unique(EMPI)]; length(copd_empi)
asthma[, .N, .(copd_existence, EMPI %in% copd_empi)][order(copd_existence, EMPI)]
asthma[, copd_alldef := ifelse(copd_existence == "Yes" | EMPI %in% copd_empi, 1, 0)]
# To be prudent, count all definitions of COPD as having the dx (coded as 1)

```

## 2.2. Alpha-1-antitrypsin deficiency (leads to early onset COPD)

```{r}

a1ad_nm <- dia_condensed[grepl("(?i)antitrypsin", Diagnosis_Name), unique(Diagnosis_Name)]
a1ad_nm
a1ad_empi <- dia[Diagnosis_Name %in% a1ad_nm, unique(EMPI)]; length(a1ad_empi)
asthma[, .N, .(EMPI %in% a1ad_empi, copd_alldef)][order(EMPI, copd_alldef)]
asthma[, a1ad := ifelse(EMPI %in% a1ad_empi, 1, 0)]

```

## 2.3. Cystic fibrosis

```{r}

cf_nm <- dia_condensed[grepl("(?i)cystic fibrosis", Diagnosis_Name), ][!grepl("(?i)screening|carrier", Diagnosis_Name), 
                                                                       unique(Diagnosis_Name)]
cf_nm
cf_empi <- dia[Diagnosis_Name %in% cf_nm, unique(EMPI)]; length(cf_empi)
asthma[, .N, .(EMPI %in% cf_empi, copd_alldef)][order(EMPI, copd_alldef)]
asthma[, cf := ifelse(EMPI %in% cf_empi, 1, 0)]

```

## 2.4. Bronchiectasis

```{r}

bronchiectasis_nm <- dia_condensed[grepl("(?i)bronchiectasis", Diagnosis_Name), unique(Diagnosis_Name)]
bronchiectasis_nm
bronchiectasis_empi <- dia[Diagnosis_Name %in% bronchiectasis_nm, unique(EMPI)]; length(bronchiectasis_empi)
asthma[, bronchiectasis := ifelse(EMPI %in% bronchiectasis_empi, 1, 0)]

```

## 2.5. Chronic bronchitis

```{r}

cbronchitis_nm <- dia_condensed[grepl("(?i)chronic bronchitis", Diagnosis_Name), unique(Diagnosis_Name)]
cbronchitis_nm
cbronchitis_empi <- dia[Diagnosis_Name %in% cbronchitis_nm, unique(EMPI)]; length(cbronchitis_empi)
asthma[, cbronchitis := ifelse(EMPI %in% cbronchitis_empi, 1, 0)]

```

## 2.6. Immune deficiencies

```{r}

immunodef_nm <- dia_condensed[grepl("(?i)immunodeficiency", Diagnosis_Name), ][!grepl("(?i)exposure|screening|family history", 
                                                                                      Diagnosis_Name), unique(Diagnosis_Name)]
immunodef_nm
immunodef_empi <- dia[Diagnosis_Name %in% immunodef_nm, unique(EMPI)]; length(immunodef_empi)
asthma[, immunodef := ifelse(EMPI %in% immunodef_empi, 1, 0)]

asthma[copd_alldef == 1 | a1ad == 1 | cf == 1 | bronchiectasis == 1 | cbronchitis == 1, ] %>% nrow()

```

# 2.7. Pulmonary hypertension

```{r}

pulhtn_nm <- dia_condensed[grepl("(?i)pulmonary", Diagnosis_Name), ][grepl("(?i)hypertension", Diagnosis_Name), 
                                                                     unique(Diagnosis_Name)]
pulhtn_nm
pulhtn_empi <- dia[Diagnosis_Name %in% pulhtn_nm, unique(EMPI)]; length(pulhtn_empi)
asthma[, pulhtn := ifelse(EMPI %in% pulhtn_empi, 1, 0)]

```

## 2.8. Allergic rhinitis

```{r}

arhinitis_nm <- dia_condensed[grepl("(?i)allergic rhinitis", Diagnosis_Name), unique(Diagnosis_Name)]
arhinitis_nm
arhinitis_empi <- dia[Diagnosis_Name %in% arhinitis_nm, unique(EMPI)]; length(arhinitis_empi)
asthma[, arhinitis := ifelse(EMPI %in% arhinitis_empi, 1, 0)]

```

## 2.9. Congestive heart failure 

```{r}

chf_nm <- dia_condensed[grepl("(?i)heart failure", Diagnosis_Name), ][grepl("(?i)congestive", Diagnosis_Name), 
                                ][!grepl("(?i)without", Diagnosis_Name), unique(Diagnosis_Name)]
chf_nm
chf_empi <- dia[Diagnosis_Name %in% chf_nm, unique(EMPI)]; length(chf_empi)
asthma[, chf := ifelse(EMPI %in% chf_empi, 1, 0)]

```

## 2.10. Acid reflux

```{r}

gerd_nm <- dia_condensed[grepl("(?i)esophageal reflux|acid reflux", Diagnosis_Name), unique(Diagnosis_Name)]
gerd_nm
gerd_empi <- dia[Diagnosis_Name %in% gerd_nm, unique(EMPI)]; length(gerd_empi)
asthma[, gerd := ifelse(EMPI %in% gerd_empi, 1, 0)]

rm(dia, dia_condensed)

```

# 3. Medication data
## 3.1. Read in data & parse out corticosteroids

```{r}

med <- fread(str_c(rpdr_file_header, "Med", rpdr_file_ending))
med[, unique(EMPI)] %in% asthma[, EMPI] %>% table()
med <- med[EMPI %in% asthma$EMPI, ]
med[, Medication_Date := as.Date(Medication_Date, format = "%m/%d/%Y")]

asthma[, Collect_Date := as.Date(Last_Collection_Date, "%Y-%m-%d")]
med <- merge(med, asthma[, .(EMPI, Collect_Date)], by = "EMPI", all.x = T)
med[, med_timegap_d := as.numeric(Collect_Date - Medication_Date)]
med[, med_timegap_y := med_timegap_d/365.25]
med[, summary(med_timegap_y)]


## Mapping file for medications of interest

med_map <- read_xlsx(med_map_fname) %>% as.data.table()
med_map_cs <- med_map[Medication_GWAS_Group %in% c("Oral Corticosteroids", "Inhaled Corticosteroids", 
                                                   "Inhaled Corticosteroids and Long Acting Beta Agonists") | 
                        Medication_Pegasus_Group %in% c("Oral Steroids", "Inhaled Steroids", "Cortisone", 
                                                        "Hydrocortisone"), ]
datatable(med_map_cs[, .(Medication_Name, Medication_GWAS_Group, Medication_Pegasus_Group, 
                         ICS_or_ICS_LABA, Medication_Biobank_Folder)])

med_condensed <- med %>% group_by(Medication) %>% summarise()
med_all_subgroups <- data.frame(Medication = character(),
                                Medication_Name = character(),
                                Medication_Biobank_Folder = character(),
                                Medication_GWAS_Group = character(),
                                Medication_Pegasus_Group = character())
for (m in 1:nrow(med_map_cs)) {
        med_subSampleGroup <- med_condensed %>% filter(grepl(med_map_cs[m, ]$Search_Term,
                                                        Medication,
                                                        ignore.case = med_map_cs[m, ]$Ignore.case,
                                                        perl = med_map_cs[m, ]$Perl)) %>%
                mutate(Medication_Name  = med_map_cs[m, ]$Medication_Name,
                       Medication_Biobank_Folder = med_map_cs[m, ]$Medication_Biobank_Folder,
                       Medication_GWAS_Group = med_map_cs[m, ]$Medication_GWAS_Group,
                       Medication_Pegasus_Group = med_map_cs[m, ]$Medication_Pegasus_Group)
        med_all_subgroups <- rbind(med_all_subgroups, med_subSampleGroup)
        rm(med_subSampleGroup)
}
med_all_subgroups <- as.data.table(med_all_subgroups)
med_all_subgroups[Medication %in% med_all_subgroups[duplicated(Medication), Medication], ]
# Pulmicort flexhaler 180 mcg (budesonide inhalation powder 180 mcg) - LMR 5919 does not seem to 
# contain dexamethasone - remove record
med_all_subgroups <- med_all_subgroups[!(Medication == "Pulmicort flexhaler 180 mcg (budesonide inhalation powder 180 mcg) - LMR 5919" &
                                           Medication_Name == "Dexamethasone")]
med_all_subgroups <- med_all_subgroups[!(Medication == "Id-Fluticasone Or Fluticasone/Salmeterol (2013p002453) Inhaler" &
                                           Medication_Name == "Fluticasone/salmeterol")]
med_all_subgroups[, ':='(ocs = case_when(Medication_GWAS_Group == "Oral Corticosteroids" | 
                                           Medication_Pegasus_Group == "Oral Steroids" ~ 1, 
                                         TRUE ~ 0), 
                         ics = case_when(Medication_GWAS_Group %in% c("Inhaled Corticosteroids", 
                                                                      "Inhaled Corticosteroids and Long Acting Beta Agonists") | 
                                           Medication_Pegasus_Group == "Inhaled Steroids" ~ 1, 
                                         TRUE ~ 0), 
                         cortisone_hydro = case_when(Medication_Pegasus_Group %in% c("Cortisone", "Hydrocortisone") ~ 1, 
                                                     TRUE ~ 0))]

med_cs <- merge(med, med_all_subgroups, by = "Medication")
med_cs[, unique(EMPI)] %in% asthma[, EMPI] %>% table() 

```

## 3.2. Filter out medications prescribed after serum sample collection

This will effectively exclude subjects with no serum collection date

```{r}

med_cs_trim <- med_cs[Medication_Date <= Collect_Date, ]
med_cs_trim[, length(unique(EMPI))] 

med_cs_trim[, cs_timegap_d := as.numeric(Collect_Date - Medication_Date)]
med_cs_trim[, cs_timegap_y := cs_timegap_d/365.25]
med_cs_trim[, summary(cs_timegap_y)]
ggplot(med_cs_trim) + 
        geom_histogram(aes(cs_timegap_y), bins = 200) + 
        labs(x = "Time (years) from corticosteroid prescritption to serum collection") + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))

ggplot(med_cs_trim[!is.na(med_timegap_y), max(med_timegap_y, na.rm = T), by = EMPI]) + 
        geom_histogram(aes(V1), bins = 200) + 
        labs(x = "Maximum gap between prescription and serum collection (years) for each subject", 
             caption = "n = 6046") + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12), 
              plot.caption = element_text(size = 12))

med_cs_trim[, .N, .(cs_timegap_y <= 1, cs_timegap_y <= 3, cs_timegap_y <= 5, 
                    cs_timegap_y <= 10)][order(cs_timegap_y, cs_timegap_y.1, cs_timegap_y.2, cs_timegap_y.3)]

```

## 3.3. Create variables for number of prescriptions (total and total over a period of years)

```{r}

med_cs_smry <- med_cs_trim[, .(ics_trim_totnum = sum(ics, na.rm = T), 
                               ocs_trim_totnum = sum(ocs, na.rm = T), 
                               cortisone_hydro_trim_totnum = sum(cortisone_hydro)), by = EMPI]
med_cs_smry_1y <- med_cs_trim[cs_timegap_y <= 1, .(ics_trim_totnum_1y = sum(ics, na.rm = T), 
                                                   ocs_trim_totnum_1y = sum(ocs, na.rm = T), 
                                                   cortisone_hydro_trim_totnum_1y = sum(cortisone_hydro)), by = EMPI]
med_cs_smry <- merge(med_cs_smry, med_cs_smry_1y, by = "EMPI", all = T)

med_cs_smry_2y <- med_cs_trim[cs_timegap_y <= 2, .(ics_trim_totnum_2y = sum(ics, na.rm = T), 
                                                   ocs_trim_totnum_2y = sum(ocs, na.rm = T), 
                                                   cortisone_hydro_trim_totnum_2y = sum(cortisone_hydro)), by = EMPI]
med_cs_smry <- merge(med_cs_smry, med_cs_smry_2y, by = "EMPI", all = T)

med_cs_smry_3y <- med_cs_trim[cs_timegap_y <= 3, .(ics_trim_totnum_3y = sum(ics, na.rm = T), 
                                                   ocs_trim_totnum_3y = sum(ocs, na.rm = T), 
                                                   cortisone_hydro_trim_totnum_3y = sum(cortisone_hydro)), by = EMPI]
med_cs_smry <- merge(med_cs_smry, med_cs_smry_3y, by = "EMPI", all = T)

med_cs_smry_5y <- med_cs_trim[cs_timegap_y <= 5, .(ics_trim_totnum_5y = sum(ics, na.rm = T), 
                                                   ocs_trim_totnum_5y = sum(ocs, na.rm = T), 
                                                   cortisone_hydro_trim_totnum_5y = sum(cortisone_hydro)), by = EMPI]
med_cs_smry <- merge(med_cs_smry, med_cs_smry_5y, by = "EMPI", all = T)

med_cs_smry_10y <- med_cs_trim[cs_timegap_y <= 10, .(ics_trim_totnum_10y = sum(ics, na.rm = T), 
                                                     ocs_trim_totnum_10y = sum(ocs, na.rm = T), 
                                                     cortisone_hydro_trim_totnum_10y = sum(cortisone_hydro)), by = EMPI]
med_cs_smry <- merge(med_cs_smry, med_cs_smry_10y, by = "EMPI", all = T)


## Set NA to 0

cs_cnt_cols <- grep("_trim_totnum", colnames(med_cs_smry), value = T)
cs_cnt_cols
med_cs_smry[, (cs_cnt_cols) := lapply(.SD, function(x) {replace(x, is.na(x), 0)}), .SDcols = cs_cnt_cols]

```

## 3.4. Add an indicator for OCS prescribed within 10 days before serum collection

```{r}

# datatable(med_cs_trim[ocs == 1 & cs_timegap_d <= 10, 
#                       .(EMPI, Medication, Medication_Date, Collect_Date, ocs, cs_timegap_d)])
med_cs_trim[ocs == 1 & cs_timegap_d <= 10, length(unique(EMPI))]
med_cs_smry[, ocs_ard_serum_coll := case_when(EMPI %in% med_cs_trim[ocs == 1 & cs_timegap_d <= 10, unique(EMPI)] ~ 1, 
                                              TRUE ~ 0)]

# datatable(med_cs_trim[cortisone_hydro == 1 & cs_timegap_d <= 10, 
#                       .(EMPI, Medication, Medication_Date, Collect_Date, cortisone_hydro, cs_timegap_d)])
med_cs_smry[, cortisone_ard_serum_coll := case_when(EMPI %in% med_cs_trim[cortisone_hydro == 1 & cs_timegap_d <= 10, 
                                                                          unique(EMPI)] ~ 1, 
                                                    TRUE ~ 0)]
med_cs_smry[, .N, .(ocs_ard_serum_coll, cortisone_ard_serum_coll)]

```

## 3.5. Plot corticosteroid medication counts

```{r}

med_cs_smry[, summary(ics_trim_totnum)]
med_cs_smry[, summary(ocs_trim_totnum)]
med_cs_smry[, summary(cortisone_hydro_trim_totnum)]


## ICS

g1 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ics_trim_totnum_1y), bins = 100) + 
        labs(x = "1yr before") + 
        ylim(c(0, 3500)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g2 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ics_trim_totnum_2y), bins = 100) + 
        labs(x = "2yrs before") + 
        ylim(c(0, 3500)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g3 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ics_trim_totnum_3y), bins = 100) + 
        labs(x = "3yrs before") + 
        ylim(c(0, 3500)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g4 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ics_trim_totnum_5y), bins = 100) + 
        labs(x = "5yrs before") + 
        ylim(c(0, 3500)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g5 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ics_trim_totnum_10y), bins = 100) + 
        labs(x = "10yrs before") + 
        ylim(c(0, 3500)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g6 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ics_trim_totnum), bins = 100) + 
        labs(x = "Entire time in system") + 
        ylim(c(0, 3500)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
grid.arrange(g1, g2, g3, g4, g5, g6, nrow = 2, 
             top = textGrob("Number of ICS prescriptions before/on serum collection", 
                            gp = gpar(fontsize = 16)))


## OCS

g7 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ocs_trim_totnum_1y), bins = 100) + 
        labs(x = "1yr before") + 
        ylim(c(0, 4000)) +
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g8 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ocs_trim_totnum_2y), bins = 100) + 
        labs(x = "2yrs before") + 
        ylim(c(0, 4000)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g9 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ocs_trim_totnum_3y), bins = 100) + 
        labs(x = "3yrs before") + 
        ylim(c(0, 4000)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g10 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ocs_trim_totnum_5y), bins = 100) + 
        labs(x = "5yrs before") + 
        ylim(c(0, 4000)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g11 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ocs_trim_totnum_10y), bins = 100) + 
        labs(x = "10yrs before") + 
        ylim(c(0, 4000)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g12 <- ggplot(med_cs_smry) + 
        geom_histogram(aes(ocs_trim_totnum), bins = 100) + 
        labs(x = "Entire time in system") + 
        ylim(c(0, 4000)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
grid.arrange(g7, g8, g9, g10, g11, g12, nrow = 2, 
             top = textGrob("Number of OCS prescriptions before/on serum collection", 
                            gp = gpar(fontsize = 16)))

```

# 4. Detailed corticosteroids tabulated data
## 4.1. ICS

```{r}

med_cs_smry[, .N, ics_trim_totnum_1y][order(ics_trim_totnum_1y)] %>% datatable()
med_cs_smry[, .N, ics_trim_totnum_2y][order(ics_trim_totnum_2y)] %>% datatable()
med_cs_smry[, .N, ics_trim_totnum_3y][order(ics_trim_totnum_3y)] %>% datatable()
med_cs_smry[, .N, ics_trim_totnum_5y][order(ics_trim_totnum_5y)] %>% datatable()
med_cs_smry[, .N, ics_trim_totnum_10y][order(ics_trim_totnum_10y)] %>% datatable()
med_cs_smry[, .N, ics_trim_totnum][order(ics_trim_totnum)] %>% datatable()

```

## 4.2. OCS

```{r}

med_cs_smry[, .N, ocs_trim_totnum_1y][order(ocs_trim_totnum_1y)] %>% datatable()
med_cs_smry[, .N, ocs_trim_totnum_2y][order(ocs_trim_totnum_2y)] %>% datatable()
med_cs_smry[, .N, ocs_trim_totnum_3y][order(ocs_trim_totnum_3y)] %>% datatable()
med_cs_smry[, .N, ocs_trim_totnum_5y][order(ocs_trim_totnum_5y)] %>% datatable()
med_cs_smry[, .N, ocs_trim_totnum_10y][order(ocs_trim_totnum_10y)] %>% datatable()
med_cs_smry[, .N, ocs_trim_totnum][order(ocs_trim_totnum)] %>% datatable()

```

## 4.3. Cortisone/hydrocortisone

```{r}

med_cs_smry[, .N, cortisone_hydro_trim_totnum_1y][order(cortisone_hydro_trim_totnum_1y)] %>% datatable()
med_cs_smry[, .N, cortisone_hydro_trim_totnum_2y][order(cortisone_hydro_trim_totnum_2y)] %>% datatable()
med_cs_smry[, .N, cortisone_hydro_trim_totnum_3y][order(cortisone_hydro_trim_totnum_3y)] %>% datatable()
med_cs_smry[, .N, cortisone_hydro_trim_totnum_5y][order(cortisone_hydro_trim_totnum_5y)] %>% datatable()
med_cs_smry[, .N, cortisone_hydro_trim_totnum_10y][order(cortisone_hydro_trim_totnum_10y)] %>% datatable()
med_cs_smry[, .N, cortisone_hydro_trim_totnum][order(cortisone_hydro_trim_totnum)] %>% datatable()

```

# 5. Session info

```{r}

sessionInfo()

```
